move stuff and add gosum
